Introduce Routing.js file to contain routing from Page.js
